Send resource load notifications from plugins to devtools agent. This is a part of a fix which would allow to inspect plugin resources.
BUG=2084

Review URL: http://codereview.chromium.org/435044

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@33672 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/plugin/webplugin_proxy.h
@@ -48,7 +48,7 @@ class WebPluginProxy : public webkit_glue::WebPlugin {
   void SetWindowlessPumpEvent(HANDLE pump_messages_event);
 #endif
 
-  void CancelResource(int id);
+  void CancelResource(unsigned long id);
   void Invalidate();
   void InvalidateRect(const gfx::Rect& rect);
   NPObject* GetWindowScriptNPObject();
@@ -127,7 +127,7 @@ class WebPluginProxy : public webkit_glue::WebPlugin {
                                 bool notify_needed,
                                 intptr_t notify_data);
 
-  void SetDeferResourceLoading(int resource_id, bool defer);
+  void SetDeferResourceLoading(unsigned long resource_id, bool defer);
 
   bool IsOffTheRecord();
