+++ webkit/glue/plugins/plugin_stream_url.cc
@@ -13,7 +13,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace NPAPI {
 
 PluginStreamUrl::PluginStreamUrl(
-    int resource_id,
+    unsigned long resource_id,
     const GURL &url,
     PluginInstance *instance,
     bool notify_needed,
